% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyFilters.R
\name{applyFilters}
\alias{applyFilters}
\title{\code{applyFilters} - Apply filters to occupancy model outputs.}
\usage{
applyFilters(roster)
}
\arguments{
\item{roster}{String. A dataframe with columns: datPath, modPath, ver, indicator, region,
nSamps, minObs, write, outPath, clipBy, group (see \code{createRoster}).}
}
\value{
A dataframe with processed model outputs to be passed to \code{calcMSI}.
}
\description{
This function can be used to subset the occupancy model
             outputs based on number of records, priority/ pollinator status
             and region. It works at the level of the taxonomic group so
             must be applied across multiple groups if needed.
}
