% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applySamp.R
\name{applySamp}
\alias{applySamp}
\title{\code{applySamp} - Reads a series of occupancy model and summarises the outputs}
\usage{
applySamp(roster, parallel = TRUE, sample = TRUE)
}
\arguments{
\item{roster}{list}

\item{parallel}{Logical}

\item{sample}{Logical. Should the model sample from the posterior distibution or just get the a parameters instead?}
}
\description{
\code{applySamp} - Reads a series of occupancy model and summarises the outputs
}
