\docType{data}
\name{boston}
\alias{boston}
\title{Boston Real estate data}
\format{\describe{
\item{CRIM}{Crim rate by unit}
\item{NOX2}{Squared Concentration of nitrogen oxide by unit}
\item{RM2}{Squared mean number of rooms by unit}
\item{logDIS}{Logarithm of the weighted distance to 5 activity centers}
\item{PTRATIO}{Ratio between the number of student and teachers}
\item{logMEDV}{Logarithm of the median value of real estates in an unit}
}}
\usage{
boston
}
\description{
Information by units about real estates and the unit.
}
\details{
This dataset was taken from the StatLib library which is maintained at Carnegie Mellon University.
}
\author{
Harrison, D. and Rubinfeld, D.L.
}
\keyword{datasets}
