% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMSI.R
\name{calcMSI}
\alias{calcMSI}
\title{\code{calcMSI} - Calculate multispecies indicator.}
\usage{
calcMSI(dat, method, write, outPath, plotLabel, bmaInd = NULL, ...)
}
\arguments{
\item{dat}{String. Object returned by \code{applyFilters}.}

\item{method}{String. Which indicator method to use. One of "lambda" or "BMA".}

\item{write}{Logical. Whether or not to write the outputs to  file.}

\item{outPath}{String. Where to store the outputs if write = TRUE.}

\item{plotLabel}{String. Title for the indicator plot.}

\item{bmaInd}{String. Use "prime" if all species data are available for all species in all years, and 
"reg" otherwise. See Freeman et al. 2020 for details.}

\item{...}{String. Title for the indicator plot.Additional arguments to be passed to \code{BRCindicators::bma}}
}
\value{
An list with elements indicator, short term assessment, long term assessment and plot.
}
\description{
This function takes the output of \code{applyFilters} and produces a multispecies indicator
             using the chosen method.
}
