% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineSamps.R
\name{combineSamps}
\alias{combineSamps}
\title{\code{combineSamps} - function inside \code{tempSampPost}, see \code{tempSampPost} and \code{applySamp} for details}
\usage{
combineSamps(species, minObs, region, sample_n)
}
\description{
\code{combineSamps} - function inside \code{tempSampPost}, see \code{tempSampPost} and \code{applySamp} for details
}
