% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRoster.R
\name{createRoster}
\alias{createRoster}
\title{\code{createRoster} - Specify how to filter the occupancy model outputs.}
\usage{
createRoster(
  index,
  modPath = "/data-s3/occmods/",
  metaPath,
  ver = "most_recent",
  group,
  indicator,
  region,
  nSamps = 1000,
  minObs = 50,
  write,
  outPath,
  clipBy = "species",
  t0,
  tn
)
}
\arguments{
\item{index}{NO IDEA WHAT THIS IS FOR}

\item{modPath}{String. Location of the occupancy model outputs.}

\item{metaPath}{String.}

\item{ver}{String. Which set of occupancy model outputs? E.g. Charlie's are "2017_Charlie".}

\item{group}{SOMETHING}

\item{indicator}{String. Whether or not to subset species and, if so, based
on what. Options are: "priority" for priority species; 
"pollinators" for pollinators; and all to return all species in the group.}

\item{region}{String. One of "UK", "GB", "England", "Wales", "Scotland", 
or "NorthernIreland".}

\item{nSamps}{Numeric. Number of samples to extract from each species' 
posterior distribution.}

\item{minObs}{Numeric. Threshold number of observation below which a 
species is dropped from the sample.}

\item{write}{Logical. If TRUE then the outputs are written as a 
.rdata file to outPath.}

\item{outPath}{String. Location to store the outputs if write = TRUE.}

\item{clipBy}{String. One of "species" or "group" indicating whether to clip outputs by the first and last years of data for
each species or for the whole group, respectively.}
}
\value{
A list of 1-row dataframes containing all arguments needed for applyFilters. applyFilters
        can then be applied to this list to filter models outputs for different groups, which may 
        come from different rounds (e.g. Charlie's or later), for different regions, etc.
}
\description{
This function can be used to specify the filters that will be applied to 
             the occupancy model outputs using \code{applyFilters}. It works by creating a list
             of 1-row dataframes with all the information needed for \code{applyFilters}. This list 
             is then applied to the applyFilters function later. Arguments should be provided
             as vectors of equal length, with each element in the vectors corresponding to one 
             call to \code{applyFilters}.
}
