\docType{data}
\name{diabete}
\alias{diabete}
\title{Pima Indians Diabetes}
\format{\describe{
\item{Pregnancies}{Number of times pregnant}
\item{Glucose}{Plasma glucose concentration a 2 hours in an oral glucose tolerance test}
\item{BloodPressure}{Diastolic blood pressure (mm Hg)}
\item{SkinThickness}{Triceps skin fold thickness (mm)}
\item{Insulin}{2-Hour serum insulin (mu U/ml)}
\item{BMI}{Body mass index (weight in kg/(height in m)^2)}
\item{DiabetesPedigreeFunction}{Diabetes pedigree function}
\item{Age}{Age (years)}
\item{Outcome}{Class variable (0 or 1)}
}}
\usage{
diabete
}
\description{
Several constraints were placed on the selection of these instances from a larger database. In particular, all patients here are females at least 21 years old of Pima Indian heritage.
}
\author{
Smith, J.W., Everhart, J.E., Dickson, W.C., Knowler, W.C., & Johannes, R.S. (1988). Using the ADAP learning algorithm to forecast the onset of diabetes mellitus. In Proceedings of the Symposium on Computer Applications and Medical Care (pp. 261--265). IEEE Computer Society Press.
}
\keyword{datasets}

