% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMetadata.R
\name{extractMetadata}
\alias{extractMetadata}
\title{\code{extractMetadata} - Extract metadata from master table created with createMetadata}
\usage{
extractMetadata(meta)
}
\arguments{
\item{meta}{table created by createMetadata}
}
\value{
Dataframe of metadata
}
\description{
This function extract more metadata from input data and model output files
}
