% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA.R
\name{getA}
\alias{getA}
\title{\code{applyFilters} - Extracts the "a" parameter from occupancy model outputs.}
\usage{
getA(
  indata = "../data/model_runs/",
  keep,
  REGION_IN_Q = "a",
  group_name = "",
  combined_output = TRUE,
  max_year_model = NULL,
  min_year_model = NULL,
  write = FALSE,
  minObs = NULL,
  t0,
  tn,
  parallel = TRUE,
  n.cores = NULL
)
}
\description{
"a" is the occupancy on the logit scale
Currently this only works for Regions. For the whole domain some recoding and calculation would be required.
This code has been copied from tempSampPost(). There is potential redundancy that could be streamlined at a later date
The data extracted this way are what we need for the bma method
}
