\docType{data}
\name{ozone}
\alias{ozone}
\title{Ozone Concentration}
\format{\describe{
\item{T12}{Temperature at 12h}
\item{T15}{Temperature at 15h}
\item{Vx}{Wind}
\item{Ne12}{Nebulosity at 12h}
}}
\usage{
ozone
}
\description{
This data was extracted from the 1994 Census bureau database by Ronny Kohavi and Barry Becker (Data Mining and Visualization, Silicon Graphics). A set of reasonably clean records was extracted using the following conditions: ((AAGE>16) && (AGI>100) && (AFNLWGT>1) && (HRSWK>0)). The prediction task is to determine whether a person makes over $50K a year.
}
\author{
Ronny Kohavi and Barry Becker
}
\keyword{datasets}

