% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpeciesTrend.R
\name{plotSpeciesTrend}
\alias{plotSpeciesTrend}
\title{\code{plotSpeciesTrend} - Plot one species' trend, with 95% credible intervals}
\usage{
plotSpeciesTrend(dat, species)
}
\arguments{
\item{dat}{String. The samp_post element of the \code{applySamp} output}

\item{pspecies}{Name of the species to plot}
}
\description{
\code{plotSpeciesTrend} - Plot one species' trend, with 95% credible intervals
}
