% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseMSI.R
\name{summariseMSI}
\alias{summariseMSI}
\title{\code{summariseMSI} - Summarise the multispecies indicator for reporting to e.g. JNCC.}
\usage{
summariseMSI(minYear, maxYear, label, plotType, indicator, method)
}
\arguments{
\item{minYear}{Numeric. First year of the indicator.}

\item{maxYear}{Numeric. Final year of the indicator.}

\item{label}{String. title for indicator plot.}

\item{plotType}{String. Type of plot to produce. Options are "indicator" to plot the indicator, or "nSpecies"
to plot the number of species contributing to each year of the indicator.}

\item{indicator}{String. Object returned by \code{calcMSI}.}

\item{method}{String. Method used to calculate the indicator. One of "lambda" or "bma".}
}
\value{
If method = "lambda" returns summary stats for reporting to JNCC. Otherwise produces a plot of he indicator.
}
\description{
This function takes the output of \code{calcMSI} and summarises the outputs either as a plot or in the 
             format JNCC require for the UK biodiversity indicators.
}
