% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempSampPost.R
\name{tempSampPost}
\alias{tempSampPost}
\title{\code{tempSampPost} - from TrendSummaries, edited}
\usage{
tempSampPost(
  indata = "../data/model_runs/",
  keep,
  keep_iter,
  output_path = "../data/sampled_posterior_1000/",
  region,
  sample_n = 999,
  tolerance = 0,
  group_name = "",
  combined_output = TRUE,
  max_year_model = NULL,
  min_year_model = NULL,
  write = FALSE,
  minObs = NULL,
  scaleObs = "global",
  t0,
  tn,
  parallel = TRUE,
  n.cores = NULL,
  filetype = "rdata"
)
}
\description{
\code{tempSampPost} - from TrendSummaries, edited
}
