% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempStackFilter.R
\name{tempStackFilter}
\alias{tempStackFilter}
\title{\code{tempStackFilter} - Version of StackFilter from TrendSummaries, edited}
\usage{
tempStackFilter(
  input = "memory",
  dat = NULL,
  indata = NULL,
  output_path,
  group_name,
  metadata,
  region = NULL,
  minObs = NULL,
  maxStartGap = NULL,
  maxEndGap = NULL,
  maxMiddleGap = NULL,
  keepSpecies = NULL,
  removeSpecies = NULL,
  ClipFirst = FALSE,
  ClipLast = FALSE,
  write = FALSE
)
}
\description{
\code{tempStackFilter} - Version of StackFilter from TrendSummaries, edited
}
