% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testRoster.R
\docType{data}
\name{testRoster}
\alias{testRoster}
\title{Example roster output}
\format{
An object of class \code{"list"}
}
\usage{
data(testRoster)
}
\description{
Used for testing
}
\examples{
\dontrun{
data(testRoster)
filterDat <- bind_rows(lapply(testRoster, applyFilters))
}
}
\keyword{datasets}
