% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyClipping.R
\name{applyClipping}
\alias{applyClipping}
\title{\code{applyClipping} - Apply filters to occupancy model outputs.}
\usage{
applyClipping(data, parallel = TRUE, clipBy = "group")
}
\arguments{
\item{parallel}{Boolean. Should the operatoin run in parallel? If so then use n.cores-1.}

\item{clipBy}{String. Must be either Species or Group}

\item{roster}{String. A dataframe with columns: datPath, modPath, ver, indicator, region,
nSamps, minObs, write, outPath, clipBy, group (see \code{createRoster}).}

\item{meta}{metadata produced by applySamp}
}
\value{
A dataframe with processed model outputs to be passed to \code{calcMSI}.
}
\description{
This function can be used to subset the occupancy model
             outputs based on number of records, priority/ pollinator status
             and region. It works at the level of the taxonomic group so
             must be applied across multiple groups if needed.
}
