% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineSamps.R
\name{combineSamps}
\alias{combineSamps}
\title{\code{combineSamps} - function inside \code{tempSampPost}, see \code{tempSampPost} and \code{applySamp} for details}
\usage{
combineSamps(
  species,
  indata,
  keep_iter,
  region,
  sample_n,
  tolerance,
  combined_output,
  minObs,
  scaleObs,
  t0,
  tn,
  filetype,
  iter
)
}
\description{
\code{combineSamps} - function inside \code{tempSampPost}, see \code{tempSampPost} and \code{applySamp} for details
}
