% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMetadata.r
\name{createMetadata}
\alias{createMetadata}
\title{\code{createMetadata} - Create master metadata table from objectstore data
repository}
\usage{
createMetadata(file_location)
}
\arguments{
\item{file_location}{Path to root directory of occupancy models on object
store}
}
\value{
Dataframe of metadata
}
\description{
This function create
}
