% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testFilterDat.R
\docType{data}
\name{testFilterDat}
\alias{testFilterDat}
\title{Example applyFilters output}
\format{
An object of class \code{"data.frame"}
}
\usage{
data(testFilterDat)
}
\description{
Used for testing
}
\examples{
\dontrun{
data(testFilterDat)
MSI <- calcMSI(testFilterDat, method = "lambda")
}
}
\keyword{datasets}
